Litt rydding, klargjøre for parp

+++ app/src/main/java/no/uio/ifi/in2000/adrianch/adrianch/skumring/model/locationforecast/Geometry.kt
@@ -1,6 +1,6 @@
 package no.uio.ifi.in2000.adrianch.adrianch.skumring.model.locationforecast
 
 data class Geometry(
-    val coordinates: List<Int>,
+    val coordinates: List<Float>,
     val type: String
 )
